+++ Modern JavaScript From The Beginning/Section 8: API Projects/73 - 77. Github Finder/app.js
@@ -1,1 +1,27 @@
+const searchInput = document.getElementById('searchInput');
 
+const API = new GitHubAPI();
+
+searchInput.addEventListener(
+    'keyup',
+    event => {
+        const username = event.target.value;
+
+        if (username) {
+            API
+                .getUser(username)
+                .then(data => {
+                    if (data.profile.message === 'Not Found') {
+                        // show error alert
+                    } else {
+                        // show profile
+                    }
+                })
+                .catch(error => {
+                    console.log(error)
+                });
+        } else {
+            // clear profile
+        }
+    }
+);