+++ Modern JavaScript From The Beginning/Section 8: API Projects/73 - 77. Github Finder/github.js
@@ -1,1 +1,13 @@
+class GitHubAPI {
+    async getUser (username) {
+        const profileResponse = await fetch(
+            `https://api.github.com/users/${username}`
+        );
 
+        const profile = await profileResponse.json();
+
+        return {
+            profile
+        };
+    }
+}